+++ packages/blob-stores/CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Removed
 
 ### Fixed
+ - `setDocument` endpoint return valid file size for dirty document, based on `fs.statSync()` result  
 
 ## [5.19.1] - 2021-02-03
 ## [5.19.0] - 2021-02-02